refactor(nextly): lift the single webhook-doc builder out of the mutation service
Publishing every language records the same outbox events the update path does,
from a second write path. Both helpers were private methods reachable only from
`update`, and each needed the service for exactly one collaborator, so they move
to a module that takes it as a parameter rather than being copied into the new
caller.
